requirements.txt
gspread==3.2.0
gspread-dataframe==3.0.4
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
mysql-connector-python==8.0.18
mysqlclient==1.4.4
pandas==1.0.0
pymongo

